+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix `Date#timeAgo()` returning big hours, minutes & seconds when date is more than a month ago
 * Make `Array#sortByPath()` handle comparing `undefined` and objects better
 * Passing `headers` option to `Request#setOptions()` will no longer unset earlier headers
+* Add `RegExp.interpretWildcard(str, flags)`
 
 ## 0.6.6 (2019-02-25)